Module Cemper

module Acoustic_interval : sig ... end

module Frequency : sig ... end

module Interval : sig ... end

module Natural_ratio : sig ... end

A representant of strictly positive ratio of two natural numbers.

module Note : sig ... end

module Roman_numeral : sig ... end

Manipulating small integers displayed as roman numerals. By convention, roman numerals are used to designate the vibrating strings of the instruments, with I (1) being the higher string, and increasing numbers going down. For example, the lower C string on the 4-strings cello is IV (4).

module Scale : sig ... end

module System : sig ... end

val hello_world : Core.Sexp.t

val main : Core.Command.t