Module Cemper.Cello

A number of utils applicable to the applications of the library to the cello.

module Fingerboard_position_name : sig ... end

val fingerboard_position : Fingerboard_position_name.t -> Fingerboard_position.t

val fingerboard_highest_note : Note.t

The selection of this particular note is somewhat arbitrary, anything around that note is realistic here. This is used to limit the generation of scales and positions.

val find_fingerboard_position_exn : System.t -> Fingerboard_position_name.t -> Fingerboard_position.t