Module Cemper.System

type t

val sexp_of_t : t -> Sexplib0.Sexp.t

val create : high_vibrating_string:Note.t -> pitch:Frequency.t -> intervals_going_down:Cemper__Characterized_interval.t array -> t

Instantiate a system with open strings starting from the highest vibrating strings, and with open strings following the given intervals while going down the instruments. The resulting system is set with 1 more string than the length of the array.

val reset_pitch : t -> Roman_numeral.t -> pitch:Frequency.t -> unit

Reset the pitch from any given string.

val pitch : t -> Cemper__Fingerboard_location.t -> Frequency.t

val acoustic_interval : t -> from:Cemper__Fingerboard_location.t -> to_:Cemper__Fingerboard_location.t -> Acoustic_interval.t option

Returns None if from has a higher pitch than to_.

val add_fingerboard_position_exn : t -> Cemper__Fingerboard_position.t -> unit

Add a new position to the system. A call to this function will cause multiple fingerboard locations to be added to the system systematically, in particular:

the locations that are at the same distance to the open string on all other strings;

the locations on the same string at higher octaves.

The name must be unique and not previously known by that system. Because positions are systematically added at the octaves of the given string portion, the one given as reference for a new position is expected to be lower than one octave to the open string.

val fingerboard_positions : t -> Cemper__Fingerboard_position.t list

Returns the list of fingerboard_positions known by this system, retuned in the order they are found going up the fingerboard starting from the open string.

val find_fingerboard_position_exn : t -> name:string -> Cemper__Fingerboard_position.t

Find the position created with the given name if present or raise.